reorganising record tag input code
