+++ src/app/components/RecordTagInput/base.less
@@ -1,0 +1,5 @@
+.record-tag-input {
+  .ui.search .prompt {
+    border-radius: 4px;
+  }
+}
