+++ src/app/components/RecordTagInput/index.tsx
@@ -1,84 +1,96 @@
 import * as React from 'react';
-import {Search, Icon} from 'semantic-ui-react';
-import {TagInput} from 'components/TagInput';
+import {Icon, Label} from 'semantic-ui-react';
+import {RecordTagInput} from './base';
+import {suggestTags, ISuggestTagsResult, suggestQuestionnaireTags} from 'apollo/modules/tags';
+import {renderArray} from '../../helpers/react';
+import {isNullOrUndefined} from 'util';
 import './style.less';
-import {ITagResult, getTags} from 'apollo/modules/tags';
 
 interface IProps {
-  onChange: (tags: string[]) => void;
-  data?: ITagResult;
+  data?: ISuggestTagsResult;
+  id: string;
+
+  // pass through to RecordTagInput
   allowNewTags?: boolean;
+  onChange: (tags: string[]) => void;
   tags: string[];
-  id?: string;
+  inputID?: string;
 }
 
-function getMatchingTags(systemTags: string[], selectedTags: string[], q: string) {
-  return systemTags
-    .filter((t) => selectedTags.indexOf(t) === -1) // don't show tags that have already been selected
-    .filter((t) => t.toLowerCase().includes(q.toLowerCase()));
+function filterAlreadySelectedTags(suggested: string[], selected: string[]) {
+  return suggested.filter((t) => selected.indexOf(t) === -1);
 }
 
-const newTagTitle = 'Create a new tag';
-const newTagDesc = 'Click here or the plus icon';
-
-class RecordTagInputInner extends React.Component<IProps, any> {
+class RecordTagInputWithSuggestionsInner extends React.Component<IProps, any> {
 
   constructor(props) {
     super(props);
-    this.renderInput = this.renderInput.bind(this);
+    this.renderTag = this.renderTag.bind(this);
+    this.renderSuggested = this.renderSuggested.bind(this);
+    this.addTag = this.addTag.bind(this);
   }
 
-  private renderInput(props) {
-    const {onChange, value, addTag, ...other} = props;
-    const addTagWithVal = () => addTag(value);
-    const handleResultSelect = (_, { result }) => {
-      if (result.title === newTagTitle) {
-        return addTagWithVal();
-      }
-      addTag(result.title);
-    };
-    const results: any[] = getMatchingTags(this.props.data.getTags || [], this.props.tags, value)
-      .map((t) => ({title: t}));
-    let icon;
-    if (this.props.allowNewTags !== false) {
-      results.push({
-        title: newTagTitle,
-        description: newTagDesc,
-      });
-      icon = (<Icon name="add" link={true} onClick={addTagWithVal} />);
-    }
+  private addTag(t: string): () => void {
+    return () => this.props.onChange(this.props.tags.concat(t));
+  }
 
-    return (
-      <Search
-        loading={this.props.data.loading}
-        onResultSelect={handleResultSelect}
-        onSearchChange={onChange}
-        results={results}
-        open={value.length > 0}
-        fluid={true}
-        value={value}
-        icon={icon}
-        noResultsMessage={'No matching tags found'}
-        {...other}
-      />
+  private renderTag(tag: string): JSX.Element {
+    return(
+      <Label as="a" key={tag} onClick={this.addTag(tag)}>
+        <span>{tag}</span>
+        <Icon name="plus"/>
+      </Label>
     );
+  }
+
+  private renderSuggested(props: IProps): JSX.Element {
+    const wrap = (inner: JSX.Element): JSX.Element => {
+      return ((
+        <div>
+          <h4 className="label suggestions">Suggested Tags</h4>
+          {inner}
+        </div>
+      ));
+    };
+    if (isNullOrUndefined(props.data)) {
+      return wrap(<span>Suggested tags will appear here</span>);
+    }
+    if (isNullOrUndefined(props.data.error) === false) {
+      return wrap(<span>Failed to load suggested tags - sorry</span>);
+    }
+    if (props.data.loading) {
+      return wrap(<span>Loading...</span>);
+    }
+    if (props.data.suggestedTags.length === 0) {
+      return wrap(<span>No suggested tags available for this beneficiary</span>);
+    }
+    const unselectedSuggestedTags = filterAlreadySelectedTags(props.data.suggestedTags, props.tags);
+    if (unselectedSuggestedTags.length === 0) {
+      return wrap(<span>All suggested tags selected</span>);
+    }
+    return wrap((
+      <span>
+      {renderArray<string>(this.renderTag, unselectedSuggestedTags)}
+    </span>
+    ));
   }
 
   public render() {
     return (
-      <div className="record-tag-input">
-        <TagInput
-          id={this.props.id}
+      <div className="record-tag-input-w-suggestions">
+        <RecordTagInput
           onChange={this.props.onChange}
-          renderInput={this.renderInput}
-          addOnBlur={false}
-          addOnKeyboard={false}
           tags={this.props.tags}
+          allowNewTags={this.props.allowNewTags}
+          id={this.props.inputID}
         />
+        {this.props.children}
+        {this.renderSuggested(this.props)}
       </div>
     );
   }
 }
 
-const RecordTagInput = getTags<IProps>(RecordTagInputInner);
-export { RecordTagInput };
+const getID = (p: IProps) => p.id;
+export const RecordTagInputWithBenSuggestions = suggestTags<IProps>(getID)(RecordTagInputWithSuggestionsInner);
+export const RecordTagInputWithQuestionnaireSuggestions = suggestQuestionnaireTags<IProps>(getID)(RecordTagInputWithSuggestionsInner);
