+++ src/app/components/RecordTagInput/style.less
@@ -1,5 +1,6 @@
-.record-tag-input {
-  .ui.search .prompt {
-    border-radius: 4px;
+.record-tag-input-w-suggestions {
+  h4.label.suggestions {
+    margin-top: 25px;
+    margin-bottom: 10px;
   }
 }
