--- src/app/components/RecordTagInputWithSuggestions/style.less
@@ -1,6 +1,0 @@
-.record-tag-input-w-suggestions {
-  h4.label.suggestions {
-    margin-top: 25px;
-    margin-bottom: 10px;
-  }
-}
